requirements.txt
streamlit>=1.28.0
requests>=2.31.0
plotly>=5.17.0
pandas>=2.1.0
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-community>=0.0.38
langchain-groq>=0.0.1
langchain-google-genai>=0.0.5
chromadb>=0.4.22
streamlit-extras>=0.3.6
Pillow>=10.0.0
numpy>=1.24.0
streamlit-lottie>=0.0.5
